# back/openapi.json
{"openapi":"3.0.0","info":{"title":"API de Agendamento","version":"1.0.0","description":"Documentação da API do sistema de agendamento"},"paths":{"/analytics":{"get":{"summary":"Retorna estatísticas do sistema","tags":["Analytics"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Estatísticas retornadas com sucesso"}}}},"/auth/register":{"post":{"summary":"Registra um novo usuário","tags":["Auth"],"security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"responses":{"201":{"description":"Usuário registrado com sucesso"},"400":{"description":"Dados inválidos"}}}},"/auth/login":{"post":{"summary":"Realiza login do usuário","tags":["Auth"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}}}}},"responses":{"200":{"description":"Login realizado com sucesso"},"401":{"description":"Credenciais inválidas"}}}},"/auth/delete":{"delete":{"summary":"Remove o usuário autenticado","tags":["Auth"],"security":[{"bearerAuth":[]}],"responses":{"204":{"description":"Usuário removido"},"404":{"description":"Usuário não encontrado"}}}},"/auth/profile":{"get":{"summary":"Retorna o perfil do usuário autenticado","tags":["Auth"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Perfil do usuário","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/auth/edit":{"put":{"summary":"Atualiza o perfil do usuário autenticado","tags":["Auth"],"security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"responses":{"200":{"description":"Perfil atualizado"},"404":{"description":"Usuário não encontrado"}}}},"/auth/logout":{"post":{"summary":"Realiza logout do usuário autenticado","tags":["Auth"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Logout realizado com sucesso"}}}},"/auth/verify":{"post":{"summary":"Verifica se o token JWT é válido","tags":["Auth"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Token válido","content":{"application/json":{"schema":{"type":"object","properties":{"user":{"$ref":"#/components/schemas/User"}}}}}},"401":{"description":"Token inválido ou expirado"}}}},"/booking":{"post":{"summary":"Cria uma nova reserva","tags":["Bookings"],"security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Booking"}}}},"responses":{"201":{"description":"Reserva criada com sucesso"},"400":{"description":"Dados inválidos"}}},"get":{"summary":"Lista todas as reservas","tags":["Bookings"],"responses":{"200":{"description":"Lista de reservas","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Booking"}}}}}}}},"/booking/my":{"get":{"summary":"Lista reservas do usuário autenticado","tags":["Bookings"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Reservas do usuário"}}}},"/booking/all":{"get":{"summary":"Lista todas as reservas (admin)","tags":["Bookings"],"responses":{"200":{"description":"Lista completa de reservas"}}}},"/booking/today":{"get":{"summary":"Lista reservas de hoje","tags":["Bookings"],"responses":{"200":{"description":"Reservas de hoje"}}}},"/booking/week":{"get":{"summary":"Lista reservas da semana","tags":["Bookings"],"responses":{"200":{"description":"Reservas da semana"}}}},"/booking/month":{"get":{"summary":"Lista reservas do mês","tags":["Bookings"],"responses":{"200":{"description":"Reservas do mês"}}}},"/booking/calendar":{"get":{"summary":"Lista reservas para o calendário","tags":["Bookings"],"responses":{"200":{"description":"Reservas para o calendário"}}}},"/booking/filter":{"post":{"summary":"Filtra reservas por parâmetros","tags":["Bookings"],"security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"200":{"description":"Reservas filtradas"}}}},"/booking/{id}":{"get":{"summary":"Busca uma reserva pelo ID","tags":["Bookings"],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Reserva encontrada","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Booking"}}}},"404":{"description":"Reserva não encontrada"}}},"put":{"summary":"Atualiza uma reserva","tags":["Bookings"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Booking"}}}},"responses":{"200":{"description":"Reserva atualizada"},"404":{"description":"Reserva não encontrada"}}},"delete":{"summary":"Remove uma reserva","tags":["Bookings"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"Reserva removida"},"404":{"description":"Reserva não encontrada"}}}},"/booking/search":{"get":{"summary":"Busca reservas pela descrição","tags":["Bookings"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"query","name":"description","schema":{"type":"string"},"description":"Descrição da reserva"}],"responses":{"200":{"description":"Reservas encontradas"}}}},"/members":{"post":{"summary":"Cria um novo membro","tags":["Members"],"security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Member"}}}},"responses":{"201":{"description":"Membro criado com sucesso"},"400":{"description":"Dados inválidos"}}},"get":{"summary":"Lista todos os membros","tags":["Members"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Lista de membros","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Member"}}}}}}}},"/members/analytics":{"get":{"summary":"Retorna dados analíticos dos membros","tags":["Members"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Dados analíticos"}}}},"/members/{id}":{"get":{"summary":"Busca um membro pelo ID","tags":["Members"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Membro encontrado","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Member"}}}},"404":{"description":"Membro não encontrado"}}},"put":{"summary":"Atualiza um membro","tags":["Members"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Member"}}}},"responses":{"200":{"description":"Membro atualizado"},"404":{"description":"Membro não encontrado"}}},"delete":{"summary":"Remove um membro","tags":["Members"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"Membro removido"},"404":{"description":"Membro não encontrado"}}}},"/members/search":{"post":{"summary":"Busca membros por parâmetros","tags":["Members"],"security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"200":{"description":"Membros encontrados"}}}},"/members/filter":{"post":{"summary":"Filtra membros por parâmetros","tags":["Members"],"security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"200":{"description":"Membros filtrados"}}}},"/room":{"post":{"summary":"Cria uma nova sala","tags":["Rooms"],"security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Room"}}}},"responses":{"201":{"description":"Sala criada com sucesso"},"400":{"description":"Dados inválidos"}}},"get":{"summary":"Lista todas as salas","tags":["Rooms"],"responses":{"200":{"description":"Lista de salas","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Room"}}}}}}}},"/room/search":{"get":{"summary":"Busca salas por parâmetros","tags":["Rooms"],"parameters":[{"in":"query","name":"nome","schema":{"type":"string"},"description":"Nome da sala"}],"responses":{"200":{"description":"Resultado da busca"}}}},"/room/{id}":{"delete":{"summary":"Remove uma sala","tags":["Rooms"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"Sala removida"},"404":{"description":"Sala não encontrada"}}},"get":{"summary":"Busca uma sala pelo ID","tags":["Rooms"],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Sala encontrada","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Room"}}}},"404":{"description":"Sala não encontrada"}}},"put":{"summary":"Atualiza uma sala","tags":["Rooms"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Room"}}}},"responses":{"200":{"description":"Sala atualizada"},"404":{"description":"Sala não encontrada"}}}},"/scale":{"post":{"summary":"Cria uma nova escala","tags":["Scales"],"security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Scale"}}}},"responses":{"201":{"description":"Escala criada com sucesso"},"400":{"description":"Dados inválidos"}}},"get":{"summary":"Lista todas as escalas","tags":["Scales"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Lista de escalas","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Scale"}}}}}}}},"/scale/search":{"put":{"summary":"Busca escalas por parâmetros","tags":["Scales"],"security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"200":{"description":"Escalas encontradas"}}}},"/scale/duplicate/{id}":{"post":{"summary":"Duplica uma escala","tags":["Scales"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}}],"responses":{"201":{"description":"Escala duplicada"},"404":{"description":"Escala não encontrada"}}}},"/scale/{id}":{"delete":{"summary":"Remove uma escala","tags":["Scales"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"Escala removida"},"404":{"description":"Escala não encontrada"}}},"put":{"summary":"Atualiza uma escala","tags":["Scales"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Scale"}}}},"responses":{"200":{"description":"Escala atualizada"},"404":{"description":"Escala não encontrada"}}},"get":{"summary":"Busca uma escala pelo ID","tags":["Scales"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Escala encontrada","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Scale"}}}},"404":{"description":"Escala não encontrada"}}}},"/user":{"post":{"summary":"Cria um novo usuário","tags":["Users"],"security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"responses":{"201":{"description":"Usuário criado com sucesso"},"400":{"description":"Dados inválidos"}}},"get":{"summary":"Lista todos os usuários","tags":["Users"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Lista de usuários","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}}}},"/user/scale":{"get":{"summary":"Lista usuários por escala","tags":["Users"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Lista de usuários por escala"}}}},"/user/{id}":{"get":{"summary":"Busca um usuário pelo ID","tags":["Users"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Usuário encontrado","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"404":{"description":"Usuário não encontrado"}}},"delete":{"summary":"Remove um usuário","tags":["Users"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"Usuário removido"},"404":{"description":"Usuário não encontrado"}}},"put":{"summary":"Atualiza um usuário","tags":["Users"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"responses":{"200":{"description":"Usuário atualizado"},"404":{"description":"Usuário não encontrado"}}}},"/user/{id}/reset-password":{"patch":{"summary":"Reseta a senha do usuário","tags":["Users"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"password":{"type":"string"}}}}}},"responses":{"200":{"description":"Senha resetada com sucesso"},"404":{"description":"Usuário não encontrado"}}}}},"components":{"schemas":{"Booking":{"type":"object","properties":{"id":{"type":"string"},"description":{"type":"string"},"roomId":{"type":"string"},"date":{"type":"string","format":"date"},"startTime":{"type":"string"},"endTime":{"type":"string"},"repeat":{"type":"string","enum":["day","week","month","none"]},"dayRepeat":{"type":"integer"},"userId":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"CreateBookingInput":{"type":"object","required":["description","start_time","end_time"],"properties":{"description":{"type":"string","maxLength":255},"room":{"type":"string"},"room_id":{"type":"string"},"date":{"type":"string","format":"date"},"start_time":{"type":"string"},"end_time":{"type":"string"},"repeat":{"type":"string","enum":["day","week","month","none"]},"day_repeat":{"type":"integer"}}},"UpdateBookingInput":{"type":"object","properties":{"description":{"type":"string","maxLength":255},"room":{"type":"string"},"date":{"type":"string","format":"date"},"start_time":{"type":"string"},"end_time":{"type":"string"},"repeat":{"type":"string","enum":["day","week","month","none"]},"day_repeat":{"type":"integer"}}},"Member":{"type":"object","properties":{"id":{"type":"integer"},"fullName":{"type":"string"},"birthDate":{"type":"string","format":"date"},"gender":{"type":"string","enum":["Masculino","Feminino","Outro"]},"cpf":{"type":"string"},"rg":{"type":"string"},"phone":{"type":"string"},"email":{"type":"string"},"street":{"type":"string"},"number":{"type":"string"},"neighborhood":{"type":"string"},"city":{"type":"string"},"state":{"type":"string"},"cep":{"type":"string"},"motherName":{"type":"string"},"fatherName":{"type":"string"},"maritalStatus":{"type":"string","enum":["Solteiro","Casado","Viúvo","Divorciado"]},"hasChildren":{"type":"boolean"},"childrenCount":{"type":"integer"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"CreateMemberInput":{"type":"object","required":["full_name","birth_date","gender","phone","email"],"properties":{"full_name":{"type":"string"},"birth_date":{"type":"string","format":"date"},"gender":{"type":"string","enum":["Masculino","Feminino","Outro"]},"cpf":{"type":"string"},"rg":{"type":"string"},"phone":{"type":"string"},"email":{"type":"string"},"street":{"type":"string"},"number":{"type":"string"},"neighborhood":{"type":"string"},"city":{"type":"string"},"state":{"type":"string"},"cep":{"type":"string"},"mother_name":{"type":"string"},"father_name":{"type":"string"},"marital_status":{"type":"string","enum":["Solteiro","Casado","Viúvo","Divorciado"]},"has_children":{"type":"boolean"},"children_count":{"type":"integer"}}},"UpdateMemberInput":{"type":"object","properties":{"full_name":{"type":"string"},"birth_date":{"type":"string","format":"date"},"gender":{"type":"string","enum":["Masculino","Feminino","Outro"]},"cpf":{"type":"string"},"rg":{"type":"string"},"phone":{"type":"string"},"email":{"type":"string"},"street":{"type":"string"},"number":{"type":"string"},"neighborhood":{"type":"string"},"city":{"type":"string"},"state":{"type":"string"},"cep":{"type":"string"},"mother_name":{"type":"string"},"father_name":{"type":"string"},"marital_status":{"type":"string","enum":["Solteiro","Casado","Viúvo","Divorciado"]},"has_children":{"type":"boolean"},"children_count":{"type":"integer"}}},"Room":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"size":{"type":"integer"},"description":{"type":"string"},"exclusive":{"type":"boolean"},"status":{"type":"boolean"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"CreateRoomInput":{"type":"object","required":["name"],"properties":{"name":{"type":"string","maxLength":100},"size":{"type":"integer"},"description":{"type":"string","maxLength":500},"exclusive":{"type":"boolean"},"status":{"type":"boolean"}}},"UpdateRoomInput":{"type":"object","properties":{"name":{"type":"string","maxLength":100},"size":{"type":"integer"},"description":{"type":"string","maxLength":500},"exclusive":{"type":"boolean"},"status":{"type":"boolean"}}},"SearchRoomInput":{"type":"object","required":["name"],"properties":{"name":{"type":"string"}}},"Scale":{"type":"object","properties":{"id":{"type":"string"},"date":{"type":"string","format":"date","example":"25/12/2024"},"name":{"type":"string","example":"Escala de Natal"},"description":{"type":"string","example":"Escala especial para o Natal"},"direction":{"type":"object","properties":{"id":{"type":"string"},"full_name":{"type":"string"}}},"band":{"type":"object","properties":{"id":{"type":"string"},"full_name":{"type":"string"}}},"projection":{"type":"string","example":"João Silva"},"light":{"type":"string","example":"Maria Santos"},"transmission":{"type":"string","example":"Pedro Costa"},"camera":{"type":"string","example":"Ana Lima"},"live":{"type":"string","example":"Carlos Oliveira"},"sound":{"type":"string","example":"Fernanda Rocha"},"training_sound":{"type":"string","example":"Roberto Alves"},"photography":{"type":"string","example":"Lucia Mendes"},"stories":{"type":"string","example":"Paulo Ferreira"},"dynamic":{"type":"string","example":"Sandra Dias"}}},"CreateScaleInput":{"type":"object","required":["date","name","direction"],"properties":{"date":{"type":"string","format":"date","pattern":"DD/MM/YYYY","example":"25/12/2024","description":"Data da escala no formato DD/MM/YYYY"},"name":{"type":"string","minLength":2,"example":"Escala de Natal","description":"Nome da escala (mínimo 2 caracteres)"},"direction":{"type":"string","minLength":2,"example":"Pastor João","description":"Nome da direção (mínimo 2 caracteres)"},"band":{"type":"string","example":"Banda Jovem","description":"Nome da banda (opcional)"},"description":{"type":"string","example":"Escala especial para o Natal","description":"Descrição da escala (opcional)"},"projection":{"type":"string","example":"João Silva","description":"Responsável pela projeção (opcional)"},"light":{"type":"string","example":"Maria Santos","description":"Responsável pela iluminação (opcional)"},"transmission":{"type":"string","example":"Pedro Costa","description":"Responsável pela transmissão (opcional)"},"camera":{"type":"string","example":"Ana Lima","description":"Responsável pela câmera (opcional)"},"live":{"type":"string","example":"Carlos Oliveira","description":"Responsável pelo live (opcional)"},"sound":{"type":"string","example":"Fernanda Rocha","description":"Responsável pelo som (opcional)"},"training_sound":{"type":"string","example":"Roberto Alves","description":"Responsável pelo som de treinamento (opcional)"},"photography":{"type":"string","example":"Lucia Mendes","description":"Responsável pela fotografia (opcional)"},"stories":{"type":"string","example":"Paulo Ferreira","description":"Responsável pelas histórias (opcional)"},"dynamic":{"type":"string","example":"Sandra Dias","description":"Responsável pela dinâmica (opcional)"}}},"UpdateScaleInput":{"type":"object","properties":{"date":{"type":"string","format":"date","pattern":"DD/MM/YYYY","example":"25/12/2024","description":"Data da escala no formato DD/MM/YYYY"},"name":{"type":"string","minLength":2,"example":"Escala de Natal Atualizada","description":"Nome da escala (mínimo 2 caracteres)"},"direction":{"type":"string","minLength":2,"example":"Pastor João","description":"Nome da direção (mínimo 2 caracteres)"},"band":{"type":"string","example":"Banda Jovem","description":"Nome da banda"},"description":{"type":"string","example":"Escala especial para o Natal","description":"Descrição da escala"},"projection":{"type":"string","example":"João Silva","description":"Responsável pela projeção"},"light":{"type":"string","example":"Maria Santos","description":"Responsável pela iluminação"},"transmission":{"type":"string","example":"Pedro Costa","description":"Responsável pela transmissão"},"camera":{"type":"string","example":"Ana Lima","description":"Responsável pela câmera"},"live":{"type":"string","example":"Carlos Oliveira","description":"Responsável pelo live"},"sound":{"type":"string","example":"Fernanda Rocha","description":"Responsável pelo som"},"training_sound":{"type":"string","example":"Roberto Alves","description":"Responsável pelo som de treinamento"},"photography":{"type":"string","example":"Lucia Mendes","description":"Responsável pela fotografia"},"stories":{"type":"string","example":"Paulo Ferreira","description":"Responsável pelas histórias"},"dynamic":{"type":"string","example":"Sandra Dias","description":"Responsável pela dinâmica"}}},"SearchScaleInput":{"type":"object","required":["name"],"properties":{"name":{"type":"string","minLength":2,"example":"Natal","description":"Nome para busca (mínimo 2 caracteres)"}}},"ScaleError":{"type":"object","properties":{"error":{"type":"string","example":"Dados inválidos"},"details":{"type":"array","items":{"type":"string"},"example":["Data deve estar no formato DD/MM/YYYY.","Nome deve ter pelo menos 2 caracteres."]}}},"ScaleSuccess":{"type":"object","properties":{"message":{"type":"string","example":"Escala criada com sucesso."}}},"PaginationQuery":{"type":"object","properties":{"page":{"type":"integer","minimum":1,"default":1,"example":1,"description":"Número da página (mínimo 1)"},"pageSize":{"type":"integer","minimum":1,"maximum":100,"default":15,"example":15,"description":"Tamanho da página (1-100)"}}},"IdParam":{"type":"object","required":["id"],"properties":{"id":{"type":"string","example":"scale-123","description":"ID único da escala"}}},"User":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"phone":{"type":"string"},"email":{"type":"string"},"role":{"type":"string","enum":["user","admin"]},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"CreateUserInput":{"type":"object","required":["name","phone","email","password"],"properties":{"name":{"type":"string","maxLength":100},"phone":{"type":"string"},"email":{"type":"string"},"password":{"type":"string","minLength":6,"maxLength":50},"role":{"type":"string","enum":["user","admin"]}}},"UpdateUserInput":{"type":"object","properties":{"name":{"type":"string","maxLength":100},"phone":{"type":"string"},"email":{"type":"string"},"role":{"type":"string","enum":["user","admin"]}}},"LoginUserInput":{"type":"object","required":["email","password"],"properties":{"email":{"type":"string"},"password":{"type":"string","minLength":6,"maxLength":50}}},"ResetPasswordInput":{"type":"object","properties":{"defaultPassword":{"type":"string","minLength":6,"maxLength":50}}}}},"tags":[{"name":"Analytics","description":"Rotas de estatísticas e análises"},{"name":"Auth","description":"Rotas de autenticação e perfil"},{"name":"Bookings","description":"Gerenciamento de reservas"},{"name":"Members","description":"Gerenciamento de membros"},{"name":"Rooms","description":"Gerenciamento de salas"},{"name":"Scales","description":"Gerenciamento de escalas"},{"name":"Users","description":"Gerenciamento de usuários"}]}